# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "iKon Mall Backend",
        "description": "iKon Mall Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "locahost:5000/api",
    "basePath": "/",
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/shop": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "returns shops",
                "description": "",
                "operationId": "getShop",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "skip",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "The number of items to skip before starting to collect the result set"
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "The numbers of items to return"
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search value for shop, matches Brand, Product, Product Items name",
                        "type": "string"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category Id",
                        "type": "string"
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Current location latitude",
                        "type": "float"
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Current location longitude",
                        "type": "float"
                    },
                    {
                        "name": "time",
                        "in": "query",
                        "description": "Current time in decimal format, eg: 18:00 will be 1800",
                        "type": "integer"
                    },
                    {
                        "name": "open24hours",
                        "in": "query",
                        "description": "To check if shop opens 24/7",
                        "type": "boolean"
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "Logged-in user id",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "shops not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Shops"
                ],
                "summary": "add a new shop",
                "description": "",
                "operationId": "addShop",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Shop object that needs to be added",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Shop"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/shop/{shopId}": {
            "put": {
                "tags": [
                    "Shops"
                ],
                "summary": "Updates shop",
                "description": "",
                "operationId": "updateShop",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "path",
                        "description": "shopId that need to be updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Updated shop object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Shop"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid user supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Shops"
                ],
                "summary": "Delete shop by Id",
                "description": "",
                "operationId": "deleteShop",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "path",
                        "description": "ID of the shop that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Address not found"
                    }
                }
            },
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "Get shop data by Id",
                "description": "This API gives the data of the shop in market",
                "operationId": "getShop",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "path",
                        "description": "ID of the shop whose data is requested",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Address not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/shop/brand/{shopId}": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "Get Brands of the Shop by ShopID",
                "description": "This API gives the data of the available brandes of the products in that shop",
                "operationId": "getShop",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "path",
                        "description": "ID of the shop whose data is requested",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Address not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/shop/rating": {
            "post": {
                "tags": [
                    "Shops"
                ],
                "summary": "Rates shop",
                "description": "",
                "operationId": "rateShop",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ShopRating"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid Order"
                    }
                }
            },
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "Find shop rating by ID",
                "description": "Returns a single shop rating",
                "operationId": "getShopRatingByShopId",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "shopId",
                        "in": "query",
                        "description": "ID of shop to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Shop not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Shop": {
            "required": [
                "shopName",
                "categoryId",
                "latitude",
                "longitude",
                "vendorId"
            ],
            "properties": {
                "shopName": {
                    "type": "string",
                    "example": "Kohinoor Fashion"
                },
                "categoryId": {
                    "type": "string",
                    "example": "620257fb38c3cd30cd98bea2"
                },
                "latitude": {
                    "type": "number",
                    "example": 21.1512
                },
                "longitude": {
                    "type": "number",
                    "example": 72.7978
                },
                "vendorId": {
                    "type": "string",
                    "example": "61fe18829af8fd2eab57d7a1"
                }
            }
        },
...
        "ShopRating": {
            "required": [
                "shopId",
                "userId",
                "star"
            ],
            "properties": {
                "shopId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "star": {
                    "type": "integer",
                    "format": "int16"
                }
            }
        },
...
    }
}